+++ artemis-maven/src/main/java/com/artemis/ArtemisMaven.java
@@ -10,7 +10,6 @@ import java.util.List;
 import org.apache.maven.plugin.AbstractMojo;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugin.MojoFailureException;
-import org.apache.maven.plugin.logging.Log;
 import org.apache.maven.plugins.annotations.Component;
 import org.apache.maven.plugins.annotations.Mojo;
 import org.apache.maven.plugins.annotations.Parameter;
@@ -18,17 +17,35 @@ import org.sonatype.plexus.build.incremental.BuildContext;
 
 import com.artemis.meta.ClassMetadata;
 
+/**
+ * The artemis plugin does bytecode-weaving on annotated components.
+ */
 @Mojo(name="artemis", defaultPhase=PROCESS_CLASSES)
-public class ArtemisMaven extends AbstractMojo{
+public class ArtemisMaven extends AbstractMojo {
 
+	/**
+	 * Root folder for class files.
+	 */
 	@Parameter(property="project.build.outputDirectory")
 	private File outputDirectory;
 
+	/**
+	 * Root source folder.
+	 */
 	@Parameter(property="project.build.sourceDirectory")
 	private File sourceDirectory;
 	
+	/**
+	 * If true, will leave field stubs to keep IDE:s happy after transformations.
+	 */
 	@Parameter
 	private boolean ideFriendlyPacking;
+	
+	/**
+	 * Enabled weaving of pooled components (more viable on Android than JVM).
+	 */
+	@Parameter(defaultValue="true")
+	private boolean enablePooledWeaving;
 	
 	@Component
 	private BuildContext context;
@@ -39,13 +56,16 @@ public class ArtemisMaven extends AbstractMojo{
 		if (context != null && !context.hasDelta(sourceDirectory))
 			return;
 		
+		getLog().info(String.format("Configuration:\n\tideFriendlyPacking=%b \n\tenablePooledWeaving=%b ", 
+			ideFriendlyPacking, enablePooledWeaving));
+		
 		Weaver.retainFieldsWhenPacking(ideFriendlyPacking);
+		Weaver.enablePooledWeaving(enablePooledWeaving);
 		
 		Weaver weaver = new Weaver(outputDirectory);
 		List<ClassMetadata> processed = weaver.execute();
 		
-		Log log = getLog();
-		log.info(getSummary(processed, start));
+		getLog().info(getSummary(processed, start));
 	}
 	
 	private static CharSequence getSummary(List<ClassMetadata> processed, long start) {