+++ artemis-weaver/src/test/java/com/artemis/ComponentTypeWeaverTest.java
@@ -5,16 +5,26 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
+import org.junit.Before;
+import org.junit.Ignore;
 import org.junit.Test;
 
 import com.artemis.component.ComponentToWeave;
 import com.artemis.component.PackedToBeB;
+import com.artemis.component.PooledForced;
+import com.artemis.component.PooledNotForced;
 import com.artemis.meta.ClassMetadata;
+import com.artemis.meta.ClassMetadata.GlobalConfiguration;
 import com.artemis.meta.ClassMetadata.WeaverType;
 
 @SuppressWarnings("static-method")
 public class ComponentTypeWeaverTest {
-
+	
+	@Before
+	public void init() {
+		GlobalConfiguration.enabledPooledWeaving = true;
+	}
+	
 	@Test
 	public void pooled_weaver_test() throws Exception {
 		ClassMetadata meta = transform(ComponentToWeave.class);
@@ -31,5 +41,24 @@ public class ComponentTypeWeaverTest {
 		assertTrue(meta.foundReset); 
 		assertTrue(meta.foundEntityFor);
 		assertEquals("com/artemis/PackedComponent", meta.superClass); 
+	}
+	
+	@Test @Ignore // rewrite to match actual waeving
+	public void pooled_disbled_weaving_test() throws Exception {
+		GlobalConfiguration.enabledPooledWeaving = false;
+		
+		ClassMetadata meta = transform(PooledNotForced.class);
+		assertFalse(meta.foundReset); 
+		assertEquals("com/artemis/Component", meta.superClass); 
+	}
+	
+	@Test
+	public void pooled_forced_weaving_test() throws Exception {
+		GlobalConfiguration.enabledPooledWeaving = false;
+		
+		ClassMetadata meta = transform(PooledForced.class);
+		assertEquals(WeaverType.NONE, meta.annotation);
+		assertTrue(meta.foundReset); 
+		assertEquals("com/artemis/PooledComponent", meta.superClass); 
 	}
 }