+++ artemis-weaver/src/test/java/com/artemis/component/PooledForced.java
@@ -1,0 +1,10 @@
+package com.artemis.component;
+
+import com.artemis.annotations.PooledWeaver;
+
+
+@PooledWeaver(forceWeaving=true)
+public class PooledForced {
+	public float x;
+	public float y;
+}