+++ artemis-weaver/src/test/java/com/artemis/component/PooledNotForced.java
@@ -1,0 +1,10 @@
+package com.artemis.component;
+
+import com.artemis.annotations.PooledWeaver;
+
+
+@PooledWeaver()
+public class PooledNotForced {
+	public float x;
+	public float y;
+}