+++ artemis/src/main/java/com/artemis/annotations/PooledWeaver.java
@@ -10,4 +10,11 @@ import com.artemis.PackedComponent;
  * @see <a href="https://github.com/junkdog/artemis-odb/wiki/Component%20Types">Component types</a>
  *      on the wiki.
  */
-public @interface PooledWeaver {}
+public @interface PooledWeaver {
+	
+	/**
+	 * If true, forces weaving even if maven property <code>enablePooledWeaving</code> is
+	 * set to <code>false</code>. 
+	 */
+	boolean forceWeaving() default false;
+}