+++ artemis/src/test/java/com/artemis/MapperTest.java
@@ -1,11 +1,8 @@
 package com.artemis;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
 
-import java.lang.reflect.ParameterizedType;
-import java.lang.reflect.Type;
-
-import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;