+++ artemis/src/test/java/com/artemis/managers/GroupManagerTest.java
@@ -1,6 +1,7 @@
 package com.artemis.managers;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
 
 import org.junit.Before;
 import org.junit.Test;