+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Developed but not released
 - Direct field access works for `@PackedWeaver` components.
+- `-DideFriendlyPacking`: If true, will leave field stubs to keep IDE:s 
+  happy after transformations. Defaults to false.
+- `-DenablePooledWeaving`: Enables weaving of pooled components (more viable on
+  Android than JVM). Defaults to true.
 
 ## Version: 0.5.0 - 2013-11-24
  - Changed artemis to a multi-module project (the `artemis` folder is the old root).